feat: add highlight active section, separete components, compress readme imgs, based on #78
