+++ src/pages/ProfilePage/Favorites.tsx
@@ -1,0 +1,11 @@
+import './ProfilePage.scss';
+
+export function Favorites() {
+  return (
+    <>
+      <section className="profile__section section-my-shopping">
+        <p>Favorites section in developing...</p>
+      </section>
+    </>
+  );
+}
