+++ src/pages/ProfilePage/MyShopping.tsx
@@ -1,0 +1,11 @@
+import './ProfilePage.scss';
+
+export function MyShopping() {
+  return (
+    <>
+      <section className="profile__section section-my-shopping">
+        <p>My shopping section in developing...</p>
+      </section>
+    </>
+  );
+}
