+++ src/pages/ProfilePage/ProfilePage.scss
@@ -126,3 +126,7 @@
   width: 25px;
   height: 36px;
 }
+
+.active-section {
+  font-weight: 700;
+}
